Kernel: move i386 module to arch::i386
Start refactoring the i386 module into an architecture abstraction module. The
arch module should export functions abstracting common operations across
multiple architectures, and expose a module containing the architecture-specific
functionality.

For now, the architecture abstraction API contains two functions:

- enable_interrupts, which allow hardware to call event::dispatch_event
- disable_interrupts, which tells hardware to queue interruptions.
