+++ kernel/src/scheduler.rs
@@ -5,7 +5,7 @@ use alloc::vec::Vec;
 use core::mem;
 
 use crate::process::{ProcessStruct, ThreadStruct, ThreadState};
-use crate::i386::process_switch::process_switch;
+use crate::arch::i386::process_switch::process_switch;
 use crate::sync::{Lock, SpinLockIRQ, SpinLockIRQGuard};
 use core::sync::atomic::Ordering;
 use crate::error::{UserspaceError};
@@ -245,7 +245,7 @@ where
                 // Temporarily revive interrupts for hlt.
                 drop(interrupt_lock);
                 unsafe {
-                    crate::i386::instructions::interrupts::hlt();
+                    crate::arch::i386::instructions::interrupts::hlt();
                 }
 
                 // Kill interrupts again.
@@ -311,7 +311,7 @@ pub fn scheduler_first_schedule<F: FnOnce()>(current_thread: Arc<ThreadStruct>, 
 
     unsafe {
         // this is a new process, no SpinLockIRQ is held
-        crate::i386::instructions::interrupts::sti();
+        crate::arch::i386::instructions::interrupts::sti();
     }
 
     jump_to_entrypoint()
